Fix bug that didn't render project's todos
When a project was selected in the sidebar, the project's todos were
not rendered. This was because the sidebar was sending an outdated
project, due to it was created before any todo was add.

To fix this, now the sidebar menu sends only the project's id to the
main component, in order to always get the up-to-date info from Storage.

Also the components are now created with factory functions instead of
IIFE modules to control when the data is loaded.

Besides, now the interaction between modules is handled with event
dispatching to improve modularity.