Fix todo cards rendering
When a new todo was add, a new div#project was created inside the
existing one.

A new updateTodos function has been add to handle this correctly.

The section#todos element needs to be always present for this to work
properly (even if there's only the fallback text). A new createTodos
function has been add to manage this too.